added removeClient, check-addclient

+++ KinoSoft/KinoSoft/FormsClient/AddClient.cs
@@ -27,12 +27,13 @@ namespace KinoSoft.FormsClient
         private void button1_Click(object sender, EventArgs e)
         {
             LogicClient logicClient = new LogicClient();
+            
 
             string FirstName = textBox1.Text;
             string LastName = textBox2.Text;
             string SecondName = textBox4.Text;
             int NumberPassport = Convert.ToInt16(textBox5.Text);
-            int seriesPassport = Convert.ToInt32(textBox6.Text);
+            int seriesPassport = Convert.ToInt16(textBox6.Text);
             string PhoneNumber = textBox3.Text;
 
             logicClient.AddClient(FirstName, LastName, SecondName, NumberPassport, seriesPassport, PhoneNumber);
